Attempt to move libraries to project folder
I spent lots of time trying to incorporate all of these libraries into a single project folder, but Arduino doesn't like this. Since most of the libraries are written with references to files like this <File.h>, it will not look in our project's src/libraries folder. If I get it working, great, otherwise the default Arduino's library folder will have to suffice