Add libraries
These are copied from my Documents/Arduino/libraries folder on my laptop, from which I got a successful upload using V2.0.17 ESP32 in board manager. Not sure if everything is exactly correct, as the LCDs are not active, but the code does compile and the 7-segments work. Some modifications to these libraries may be necessary