Update both programs to use lower color depth on right display
I believe the color depth was defaulting to 16 bit, which meant that the ESP32 could not handle full screen rendering. It was limited to roughly 235x235, so there were weird cutoffs to the round display. I manually set the color depth to 8, which allows me to successfully run it at 240x240. Also adjusted the position of the sprite to spawn at 0,0 which keeps it centered